+++ scripts/constant.js
@@ -1,0 +1,3 @@
+export const MODULE_PATH = "modules/ffg-star-wars-alternative-ui";
+export const MODULE_ID = "ffg-star-wars-alternative-ui";
+export const SYSTEM_ID = "starwarsffg";